Fix scroll behavior in home when toggling grid mode
Because of the AnimatedContent in RecurringExpenseOverview
the list is reset and scrolled back to the top.
To make sure the scroll state matches that we need to reset
it here. It make the TopAppBar use the surface color again.
This is a workaround which can hopefully removed in the near
future.