Migrate the whole project to KMP
This will allow the app to be used on more platforms than just Android.
For now it works on iOS, Windows, MacOS and Linux. The publishing flow
for these platforms will be added later, for now it will still be
published for Android only. There's a known issue with the BottomSheet
not working at the moment, but this will be fixed as a follow up.