Fix next payment date calculation
Fix an issue where the date for the next payment wasn't calculated
correctly as it was using the current time of the day instead of the
day in general.

Extract days until calculation and the next payment date calculation
into a separate class and cover it with UnitTests preventing this from
happening again in the future. 🤞

fixes #272