Add basic Glance Widget
This adds a Widget using jetpack compose glance instead of xml.
This is the basic implementation updating the Widget does not work jet.
Replicating the common UpcomingPaymentsViewModel as with glance we can't
use ViewModels. Also we only have a small amount of basic composables
there is no e.g. Card with this reducing the information and style shown
in the Widget to a minimum.
The configuration of the Widget is also not yet there but prepared to be
also able to use Grid or List mode.

Signed-off-by: Dominik Dias <dias_dominik@icloud.com>