Create UnitTest for UpcomingPaymentsViewModel
Create UnitTest for the calculation of the upcoming expenses. This
verifies the critical business logic works fine and helps tracing bugs
as well as preventing them in the future.
Make the calculation independent of the current day and time of the
device in order to simulate different scenarios in the tests.